requirements.txt

azure-functions
azure-mgmt-network
azure-identity
python-jose
python-jose-cryptodome
azure-data-tables
azure-keyvault-secrets